fixed teach login issue better search component design cleaned code by using components for notification
