+++ cypress/integration/chapter/comment.spec.js
@@ -38,7 +38,8 @@ describe('Chapter comments when authenticated', () => {
 
 
         cy.get('#chapter .media-body')
-            .wait(1000) //TODO can we do better than this wait ?🤔
+            .wait(2000) //TODO can we do better than this wait ?🤔
+            .last()
             .contains(comment.text)
             .should('be.visible')
     });
