+++ frontend/app/components/main-header.hbs
@@ -43,78 +43,63 @@
       </li>
 
     </ul>
+
     <div id="myDropdown" class="dropdown">
       <div class="dropdown-content">
         <Input placeholder='Search' id="dropdownMenuButton" class="form-control search-input dropdown-toggle"
           @autocomplete="on" @click="searchLoading=false" @key-up="change" @insert-newline="enter" @bubbles="false"
           type="text" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true" />
         <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+
           {{#if this.searchLoading}}
-          <p class="dropdown-item">
-            <span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
-            {{t "messages.searching"}}
-          </p>
-          {{else}}
-          {{#each this.filterFunction as |obj|}}
           <li>
             <hr class="dropdown-divider" />
           </li>
-          <li class="dropdown-item">
-            <LinkTo @route="chapter" @model={{obj.id}} class="chapter-title">{{obj.name}}</LinkTo>
-          </li>
-          {{/each}}
+              <li class="dropdown-item">
+                <span class="spinner-border spinner-border-sm" role="status" aria-hidden="true"></span>
+                {{t "messages.searching"}}
+                </li>
+          {{else if (eq this.filterFunction.length 0)}}
+              <li>
+                <hr class="dropdown-divider" />
+              </li>
+              <li class="dropdown-item">
+                <p>{{t "main_navbar.search.result"}} <b>{{this.searchQuery}}</b> </p>
+              </li>
+          {{else}}
+              {{#each this.filterFunction as |obj|}}
+              <li>
+                <hr class="dropdown-divider" />
+              </li>
+              <div class="dropdown-item">
+
+                  <LinkTo @route="chapter" @model={{obj.id}} class="chapter-title">
+                    <div class="search">
+                      {{obj.name}}
+                    </div>
+                    <div class="chapter-metadata">
+                      <div class="d-inline mr-3">
+                      {{#if obj.verified }}
+                        {{svg-jar "verified-orange" class="verified-icon"}}
+                      {{/if}}
+                        {{svg-jar "watch" class="watch-icon"}} {{relative-time obj.createdAt}}
+                      </div>
+                      {{svg-jar "visibility" class="eye-icon"}} {{obj.views}}
+                    </div>
+                  </LinkTo>
+              </div>
+
+              {{/each}}
           {{/if}}
-
         </div>
-
       </div>
     </div>
 
     <ul class="navbar-nav ml-auto">
-
-      <BsDropdown class="nav-item notifications-dropdown d-lg-block d-none mr-4" @closeOnMenuClick={{false}} as |dd|>
-        <dd.button @type="default" class="dropdown-toggle  nav-link">
-          {{svg-jar "bell" height="27" width="27"}}
-          <span class="count">{{this.unreadNotificationsCount}}</span>
-        </dd.button>
-        <dd.menu class="dropdown-menu-lg-right" as |ddm|>
-          <a class="dropdown-item header-title" href="#">{{t "main_navbar.notifications.title"}}</a>
-          <ddm.divider class="DropdownMenuDivider"></ddm.divider>
-
-          {{#if @model.notifications}}
-          <div class="notification-items">
-            {{#each @model.notifications as |obj|}}
-            <div class="dropdown-item notification-item d-flex {{if obj.read " read" "unread" }}" href="#">
-              <div class="notification-container row" role="button" {{on "click" (fn this.onNotificationSelect obj
-                dd)}}>
-                <div class="col-12 title">
-                  {{obj.title}}
-                </div>
-                <div class="col-12 body">
-                  {{obj.body}}
-                </div>
-                <div class="col-12 time">
-                  {{relative-time obj.createdAt}}
-                </div>
-              </div>
-              <div class="notification-actions d-flex">
-                <button type="button" class="btn-delete" {{on "click" (fn this.deleteNotification obj)}}>
-                  {{svg-jar "close" width="15" height="15"}}
-                </button>
-                <button type="button" class="btn-toggle-read-status" {{on "click" (fn this.toggleNotificationStatus obj
-                  )}}>
-                  <span class="circle"></span>
-                </button>
-              </div>
-            </div>
-            {{/each}}
-          </div>
-
-          {{else}}
-          <div class="dropdown-item notification-item">{{t "main_navbar.notifications.empty"}}</div>
-          {{/if}}
-        </dd.menu>
-      </BsDropdown>
+      {{#if (eq this.router.currentRouteName "teach.login")}}
+      {{else if this.me.isAuthenticated}}
+      <Notification @model={{@model}} />
+      {{/if}}
 
     </ul>
     <ul class="navbar-nav">
