+++ frontend/app/routes/teach/index.js
@@ -6,9 +6,14 @@ export default class TeachIndexRoute extends Route {
   @service me;
 
   async model() {
-    return this.store.query('chapter', {
-      creatorId: this.me.user.id,
-      status: 'draft',
-    });
+    return {
+      chapters: await this.store.query('chapter', {
+        creatorId: this.me.user.id,
+        status: 'draft',
+      }),
+      notifications: await this.store.query('notification', {
+        recipientId: this.me.user.id,
+      }),
+    };
   }
 }
