+++ frontend/app/styles/search.scss
@@ -1,7 +1,56 @@
-.search {
-  display: flex;
+#myDropdown .dropdown-menu {
+  width: 15rem;
+  max-height: 15rem;
+  overflow: auto;
+  color: inherit;
 }
 
-.search-input {
+#myDropdown .dropdown-item {
+  white-space: normal;
+  overflow-wrap: break-word;
+  display: flex;
+  padding: .25rem 1rem;
+}
+
+#myDropdown .dropdown-item:hover,
+#myDropdown .dropdown-item:focus {
+  background-color: #e1e1e1;
+}
+
+#myDropdown .dropdown-item a {
+  color: inherit;
+  text-decoration: none;
+  display: inline-block;
+  vertical-align: middle;
+  -webkit-transform: perspective(1px) translateZ(0);
+  transform: perspective(1px) translateZ(0);
+  box-shadow: 0 0 1px rgb(0 0 0 / 0%);
+  -webkit-transition-duration: 0.3s;
+  transition-duration: 0.3s;
+  -webkit-transition-property: transform;
+  transition-property: transform;
+}
+
+#myDropdown .dropdown-item a:hover,
+#myDropdown .dropdown-item a:focus {
+  color: inherit;
+  text-decoration: none;
+}
+
+.search {
+  display: flex;
+  font-size: 3.5vh;
+}
+
+.search-input,
+.search-input:focus {
   flex: 1;
+  background-image: url(/images/icons/search.svg);
+  padding-right: calc(1.5em + .75rem);
+  background-repeat: no-repeat;
+  background-position: right calc(.375em + .1875rem) center;
+}
+
+.search-thumbnail {
+  margin-right: 10px;
 }
