+++ server/db/seeds/dev/02_chapters.js
@@ -21,6 +21,7 @@ exports.seed = function (knex) {
             description: faker.lorem.paragraph(),
             lesson_id: 'lesson1',
             content_type: 'h5p',
+            image_url: faker.image.imageUrl(328, 200, 'business', true, false),
             status: faker.random.arrayElement(status),
             content_id: `${faker.random.arrayElement(chapterId)}`,
             creator_id: faker.random.arrayElement(userIds),
@@ -41,6 +42,7 @@ exports.seed = function (knex) {
           lesson_id: 'lesson1',
           creator_id: 'user1',
           content_type: 'h5p',
+          image_url: faker.image.imageUrl(328, 200, 'business', true, false),
           content_id: 'chapter1',
           created_at: '2017-12-20 19:17:10',
           updated_at: '2017-12-20 19:17:10',
@@ -56,6 +58,7 @@ exports.seed = function (knex) {
           creator_id: 'user1',
           status: 'published',
           lesson_id: 'basics1',
+          image_url: faker.image.imageUrl(328, 200, 'business', true, false),
           content_type: 'h5p',
           content_id: 'chapter1',
           created_at: '2017-12-20 19:17:10',
@@ -72,6 +75,7 @@ exports.seed = function (knex) {
           creator_id: 'user1',
           status: 'published',
           lesson_id: 'basics2',
+          image_url: faker.image.imageUrl(328, 200, 'business', true, false),
           content_type: 'h5p',
           content_id: 'chapter1',
           created_at: '2017-12-20 19:17:10',
@@ -88,6 +92,7 @@ exports.seed = function (knex) {
           status: 'published',
           creator_id: 'user1',
           lesson_id: 'basics2',
+          image_url: faker.image.imageUrl(328, 200, 'business', true, false),
           content_type: 'h5p',
           content_id: 'chapter1',
           created_at: '2017-12-20 19:17:10',
@@ -103,6 +108,7 @@ exports.seed = function (knex) {
           creator_id: 'user1',
           description: 'qui',
           status: 'published',
+          image_url: faker.image.imageUrl(328, 200, 'business', true, false),
           lesson_id: 'basics2',
           content_type: 'h5p',
           content_id: 'chapter1',
